pyproject.toml
[project]
name = "jatayu"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.67.0",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "nats-py>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
]
